Modify Readme file and moved files to their components
