+++ src/components/Books.js
@@ -1,0 +1,32 @@
+// /* eslint-disable jsx-a11y/label-has-associated-control */
+// import React from 'react';
+
+// const Books = () => {
+//   const updateTitle = () => {
+//   };
+//   const updateAuthor = () => {
+//   };
+
+//   const submitBook = () => {
+
+//   };
+
+//   return (
+//     <div>
+//       <h3>Add new book</h3>
+//       <form>
+//         <label>Title:</label>
+//         <br />
+//         <input type="text" className="title" name="title" onChange={updateTitle} />
+//         <br />
+//         <label>Author</label>
+//         <br />
+//         <input type="text" className="author" name="author" onChange={updateAuthor} />
+//         <br />
+//         <input type="submit" value="Submit" onClick={submitBook} />
+//       </form>
+//     </div>
+//   );
+// };
+
+// export default Books;
