--- src/redux/categories/categories.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Categories() {
-  return (
-    <div>
-      <h1>Under Construction</h1>
-    </div>
-  );
-}
-
-export default Categories;
